#53 moving hasher types into their own folers
