+++ cas-dotnet-sdk/Hashers/Types/Blake2HashByteResult.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace CasDotnetSdk.Hashers.Types
+{
+    internal struct Blake2HashByteResult
+    {
+        public IntPtr result_bytes_ptr;
+        public int length;
+    }
+}
