+++ cas-dotnet-sdk/Hashers/Types/HmacSignByteResult.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace CasDotnetSdk.Hashers.Types
+{
+    internal struct HmacSignByteResult
+    {
+        public IntPtr result_bytes_ptr;
+        public int length;
+    }
+}
