+++ cas-dotnet-sdk/Hashers/Types/SHAHashByteResult.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace CasDotnetSdk.Hashers.Types
+{
+    internal struct SHAHashByteResult
+    {
+        public IntPtr result_bytes_ptr;
+        public int length;
+    }
+}
